fix(deck): 恢复模拟活动、指定队长与进阶覆盖的参数下发
worker 只读 `simulatedEvent` / `userDataOverrides` / `leaderCharacterId`
三个成组字段，页面却把它们拆成了散字段（`simType`/`simAttr`/`simUnit`、
`areaItemLevel`/`characterRank`/… ）逐个下发。散字段不会报错，只会静默失效：

- 模拟活动开关完全无效，落回真实活动分支；没选活动时抛
  `event mode requires eventId`。
- 指定队长被丢弃，队长位仍由最高技能或固定卡决定。
- 进阶数据覆盖（区域道具 / 角色等级 / 烤森门 / 玩偶加成）被丢弃，
  另有 `mysekaiGateLevelOverrides`、`mysekaiFixtureBonusRateOverrides`
  两个键名写错。

模拟面板此前只对连接世界提供加成团选择，马拉松与 5v5 只能选属性，
因此无法描述团活与混活。现在两者都能按团或按加成角色集合模拟，
VS 角色可另外指定支援团，映射到引擎的 `event_unit` 与
`custom_bonus_character_ids` / `custom_bonus_character_support_units`。
连接世界第 3 轮补上"未选章节角色"的前置校验。

两层映射同时抽成运行时零依赖的纯函数（`worker-args.ts` 表单 → worker
入参、`engine-options.ts` worker 入参 → 引擎 options），表单默认值随之
挪过去与测试共用一份，这一层才能被直接断言而不是只做源码串检查。
